Remove extra code (all of it basically)
