+++ sirdigsalot/package.xml
@@ -1,19 +1,19 @@
 <?xml version="1.0"?>
 <package>
-  <name>lk_rover</name>
+  <name>sirdigsalot</name>
   <version>0.0.0</version>
-  <description>The lk_rover package</description>
+  <description>Sirdigsalot main package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag --> 
   <!-- Example:  -->
   <!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
-  <maintainer email="brandon@todo.todo">brandon</maintainer>
+  <maintainer email="kelvin.ly1618@gmail.com">Kelvin Ly</maintainer>
 
 
   <!-- One license tag required, multiple allowed, one license per tag -->
   <!-- Commonly used license strings: -->
   <!--   BSD, MIT, Boost Software License, GPLv2, GPLv3, LGPLv2.1, LGPLv3 -->
-  <license>TODO</license>
+  <license>MIT</license>
 
 
   <!-- Url tags are optional, but mutiple are allowed, one per tag -->
@@ -43,29 +43,10 @@
   <build_depend>roscpp</build_depend>
   <build_depend>rospy</build_depend>
   <build_depend>std_msgs</build_depend>
-  <build_depend>controller_manager</build_depend>
-  <build_depend>controller_manager_msgs</build_depend>
-  <build_depend>hardware_interface</build_depend>
-  <build_depend>velocity_controllers</build_depend>
-  <build_depend>diff_drive_controller</build_depend>
-  <build_depend>message_generation</build_depend>
-  <build_depend>robot_localization</build_depend>
 
   <run_depend>roscpp</run_depend>
   <run_depend>rospy</run_depend>
   <run_depend>std_msgs</run_depend>
-  <run_depend>controller_manager</run_depend>
-  <run_depend>controller_manager_msgs</run_depend>
-  <run_depend>hardware_interface</run_depend>
-  <run_depend>diff_drive_controller</run_depend>
-  <run_depend>velocity_controllers</run_depend>
-  <run_depend>effort_controllers</run_depend>
-  <run_depend>position_controllers</run_depend>
-  <run_depend>message_runtime</run_depend>
-  <run_depend>robot_localization</run_depend>
-
-  <test_depend>gazebo_ros</test_depend>
-  <test_depend>gazebo_ros_control</test_depend>
 
   <!-- The export tag contains other, unspecified, tags -->
   <export>
